w3c-reports/12092014/mac/runned-tests/dom/nodes/Element-firstElementChild-entity-expected.txt

CONSOLE MESSAGE: line 74: TypeError: null is not an object (evaluating 'document.body.appendChild')
FAIL: Timed out waiting for notifyDone to be called
Test of Entity References
The result of this test is unknown.
